@@ -1,35 +1,24 @@
-import { useState } from "react";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
-
-
-import decoration from '../../resources/img/vision.png';
+import { MainPage, ComicsPage, ErrorPage, SingleComicPage } from "../pages";
 
 const App = () => {
-    const [charId, setCharId] = useState(null);
-    const onChangeSelectedId = (id) => {
-        setCharId(id);
-    }
+
     return (
-        <div className="app">
-            <AppHeader />
-            <main>
-                <ErrorBoundary>
-                    <RandomChar />
-                </ErrorBoundary>
-                <div className="char__content">
-                    <CharList changeSelectedId={onChangeSelectedId} />
-                    <ErrorBoundary>
-                        <CharInfo charId={charId} />
-                    </ErrorBoundary>
-                </div>
-                <img className="bg-decoration" src={decoration} alt="vision" />
-            </main>
-        </div>
+        <Router>
+            <div className="app">
+                <AppHeader />
+                <main>
+                    <Routes>
+                        <Route path="/" element={<MainPage />} />
+                        <Route path="/comics" element={<ComicsPage />} />
+                        <Route path="/comics/:comicId" element={<SingleComicPage />} />
+                        <Route path="*" element={<ErrorPage />} />
+                    </Routes>
+                </main>
+            </div>
+        </Router>
     )
 }
 
