@@ -6,9 +6,23 @@ const useMarvelService = () => {
 
     const _apiBase = 'https://gateway.marvel.com:443/v1/public/';
     const _apiKey = 'apikey=617e2d38a316c4ce64d877fad420f901';
-    const _charOffset = 210;
+    const _charsOffset = 210;
+    const _comicsOffset = 50;
 
-    const getAllCharacters = async (offset = _charOffset, limit) => {
+    const getAllComics = async (offset = _comicsOffset, limit) => {
+        const res = await request(`${_apiBase}comics?orderBy=issueNumber&limit=${limit}&offset=${offset}&${_apiKey}`);
+        return {
+            comics: res.data.results.map(_transformComic),
+            ended: res.data.total - res.data.offset <= 8
+        };
+    }
+
+    const getComic = async (id) => {
+        const res = await request(`${_apiBase}comics/${id}?${_apiKey}`);
+        return _transformComic(res.data.results[0]);
+    }
+
+    const getAllCharacters = async (offset = _charsOffset, limit) => {
         const res = await request(`${_apiBase}characters?limit=${limit}&offset=${offset}&${_apiKey}`);
         return {
             chars: res.data.results.map(_transformCharacter),
@@ -19,6 +33,18 @@ const useMarvelService = () => {
     const getCharacter = async (id) => {
         const res = await request(`${_apiBase}characters/${id}?${_apiKey}`);
         return _transformCharacter(res.data.results[0]);
+    }
+
+    const _transformComic = (comic) => {
+        let { description, thumbnail, prices, pageCount, textObjects } = comic;
+        return {
+            ...comic,
+            description: description ? description : 'There is no description',
+            thumbnail: `${thumbnail.path}.${thumbnail.extension}`,
+            price: prices[0].price ? `${prices[0].price}$` : 'Not available',
+            pages: pageCount ? `${pageCount} pages` : 'No info about pages',
+            language: textObjects[0] ? textObjects[0].language : 'en-us'
+        }
     }
 
     const _transformCharacter = (char) => {
@@ -40,7 +66,7 @@ const useMarvelService = () => {
         }
     }
 
-    return { loading, error, getAllCharacters, getCharacter }
+    return { loading, error, getAllComics, getComic, getAllCharacters, getCharacter }
 }
 
 export default useMarvelService;
